fix: 위젯, 모델화
