+++ frontend/lib/widgets/mypage/mypage_widgets.dart
@@ -2,3 +2,5 @@ export 'stat_item.dart';
 export 'list_menu.dart';
 export 'logout_modal_button.dart';
 export 'build_edit_menu.dart';
+export 'promotion_item.dart';
+export 'brand_item.dart';
